requirements.txt
pycodestyle
flask

# Config
python-dotenv

# Database
flask-sqlalchemy
flask-migrate

# Security
flask-security-too
argon2_cffi 
setuptools

flask-restful  # API handling
marshmallow  # Validation, serializing and deserialization
requests # API calls